requirements.txt
langchain
langchain-groq
langchain-community
fastapi
uvicorn
python-multipart
pypdf
ddgs
python-dotenv